requirements.txt
numpy==2.2.3
Pillow==11.1.0
PyQt6==6.8.1
PyQt6_sip==13.10.0

opencv-python~=4.11.0.86